+++ example/lib/main.dart
@@ -78,6 +78,39 @@ class _VietMapNavigationScreenState extends State<VietMapNavigationScreen> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      floatingActionButton: Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        children: [
+          FloatingActionButton(
+            onPressed: () {
+              _controller?.moveCamera(
+                  latLng: const LatLng(22.762528, 106.653099),
+                  zoom: 8,
+                  tilt: 0,
+                  bearing: 0);
+            },
+            child: const Icon(Icons.animation),
+          ),
+          FloatingActionButton(
+            onPressed: () async {
+              List<NavigationMarker>? markers =
+                  await _controller?.addImageMarkers([
+                NavigationMarker(
+                    width: 120,
+                    height: 120,
+                    imagePath: 'assets/50.png',
+                    latLng: const LatLng(10.762528, 106.653099)),
+                NavigationMarker(
+                    imagePath: 'assets/40.png',
+                    latLng: const LatLng(10.762528, 106.753099),
+                    width: 80,
+                    height: 80),
+              ]);
+            },
+            child: const Icon(Icons.animation),
+          ),
+        ],
+      ),
       body: SafeArea(
         top: false,
         child: Stack(
@@ -150,7 +183,10 @@ class _VietMapNavigationScreenState extends State<VietMapNavigationScreen> {
                 }, (r) => _showBottomSheetInfo(r));
               },
               onRouteProgressChange: (RouteProgressEvent routeProgressEvent) {
-                // print('---------------------');
+                // print('-----------ProgressChange----------');
+                // print(routeProgressEvent.currentLocation?.bearing);
+                // print(routeProgressEvent.currentLocation?.altitude);
+                // print(routeProgressEvent.currentLocation?.accuracy);
                 // print(routeProgressEvent.currentLocation?.bearing);
                 // print(routeProgressEvent.currentLocation?.latitude);
                 // print(routeProgressEvent.currentLocation?.longitude);
@@ -328,7 +364,8 @@ class _VietMapNavigationScreenState extends State<VietMapNavigationScreen> {
 
                 waypoints.add(LatLng(location.latitude, location.longitude));
                 if (data.lat != null) {
-                  waypoints.add(LatLng(data.lat ?? 0, data.lng ?? 0));
+                  waypoints.add(LatLng(
+                      data.lat?.toDouble() ?? 0, data.lng?.toDouble() ?? 0));
                 }
                 _controller?.buildRoute(waypoints: waypoints);
                 if (!mounted) return;
@@ -340,7 +377,8 @@ class _VietMapNavigationScreenState extends State<VietMapNavigationScreen> {
                 var location = await Geolocator.getCurrentPosition();
                 waypoints.add(LatLng(location.latitude, location.longitude));
                 if (data.lat != null) {
-                  waypoints.add(LatLng(data.lat ?? 0, data.lng ?? 0));
+                  waypoints.add(LatLng(
+                      data.lat?.toDouble() ?? 0, data.lng?.toDouble() ?? 0));
                 }
                 _controller?.buildAndStartNavigation(
                     waypoints: waypoints,