+++ lib/models/marker.dart
@@ -9,19 +9,28 @@ class NavigationMarker {
   final String? title;
   final String? snippet;
   int? markerId;
+  final int? width;
+  final int? height;
 
   NavigationMarker({
     required this.imagePath,
     required this.latLng,
     this.title,
     this.snippet,
-  });
+    this.width,
+    this.height,
+  })  : assert(width != null || height == null,
+            'Width and height must be both provided or both null'),
+        assert(height != null || width == null,
+            'Width and height must be both provided or both null');
   toJson() {
     return {
       'latitude': latLng.latitude,
       'longitude': latLng.longitude,
       'title': title,
-      'snippet': snippet
+      'snippet': snippet,
+      'width': width,
+      'height': height,
     };
   }
 }